cht_pdf/requirements.txt
Package                                  Version
---------------------------------------- ----------
aiohttp                                  3.9.3
aiosignal                                1.3.1
annotated-types                          0.6.0
anyio                                    4.3.0
asgiref                                  3.7.2
async-timeout                            4.0.3
attrs                                    23.2.0
backoff                                  2.2.1
bcrypt                                   4.1.2
build                                    1.0.3
cachetools                               5.3.3
certifi                                  2024.2.2
charset-normalizer                       3.3.2
chroma-hnswlib                           0.7.3
chromadb                                 0.4.24
click                                    8.1.7
coloredlogs                              15.0.1
dataclasses-json                         0.6.4
Deprecated                               1.2.14
distro                                   1.9.0
exceptiongroup                           1.2.0
faiss-cpu                                1.7.4
fastapi                                  0.110.0
filelock                                 3.13.1
flatbuffers                              23.5.26
frozenlist                               1.4.1
fsspec                                   2024.2.0
google-auth                              2.28.1
googleapis-common-protos                 1.62.0
greenlet                                 3.0.3
grpcio                                   1.62.0
h11                                      0.14.0
httpcore                                 1.0.4
httptools                                0.6.1
httpx                                    0.27.0
huggingface-hub                          0.21.3
humanfriendly                            10.0
idna                                     3.6
importlib-metadata                       6.11.0
importlib_resources                      6.1.2
jsonpatch                                1.33
jsonpointer                              2.4
kubernetes                               29.0.0
langchain                                0.1.9
langchain-community                      0.0.24
langchain-core                           0.1.27
langchain-openai                         0.0.8
langsmith                                0.1.10
marshmallow                              3.21.0
mmh3                                     4.1.0
monotonic                                1.6
mpmath                                   1.3.0
multidict                                6.0.5
mypy-extensions                          1.0.0
numpy                                    1.26.4
oauthlib                                 3.2.2
onnxruntime                              1.17.1
openai                                   1.13.3
opentelemetry-api                        1.23.0
opentelemetry-exporter-otlp-proto-common 1.23.0
opentelemetry-exporter-otlp-proto-grpc   1.23.0
opentelemetry-instrumentation            0.44b0
opentelemetry-instrumentation-asgi       0.44b0
opentelemetry-instrumentation-fastapi    0.44b0
opentelemetry-proto                      1.23.0
opentelemetry-sdk                        1.23.0
opentelemetry-semantic-conventions       0.44b0
opentelemetry-util-http                  0.44b0
orjson                                   3.9.15
overrides                                7.7.0
packaging                                23.2
pip                                      22.0.2
posthog                                  3.4.2
protobuf                                 4.25.3
pulsar-client                            3.4.0
pyasn1                                   0.5.1
pyasn1-modules                           0.3.0
pydantic                                 2.6.3
pydantic_core                            2.16.3
PyPDF2                                   3.0.1
PyPika                                   0.48.9
pyproject_hooks                          1.0.0
python-dateutil                          2.8.2
python-dotenv                            1.0.1
PyYAML                                   6.0.1
regex                                    2023.12.25
requests                                 2.31.0
requests-oauthlib                        1.3.1
rsa                                      4.9
setuptools                               59.6.0
six                                      1.16.0
sniffio                                  1.3.1
SQLAlchemy                               2.0.27
starlette                                0.36.3
sympy                                    1.12
tenacity                                 8.2.3
tiktoken                                 0.6.0
tokenizers                               0.15.2
tomli                                    2.0.1
tqdm                                     4.66.2
typer                                    0.9.0
typing_extensions                        4.10.0
typing-inspect                           0.9.0
urllib3                                  2.2.1
uvicorn                                  0.27.1
uvloop                                   0.19.0
watchfiles                               0.21.0
websocket-client                         1.7.0
websockets                               12.0
wrapt                                    1.16.0
yarl                                     1.9.4
zipp                                     3.17.0
